Bugz 2280: Bridge.sh is stupid and ugly and doesn't run right on non-ssg systems
git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@12705 219f1c6e-4a0e-43c8-96f2-6e749be13fd4

+++ installer/make_installers.pl
@@ -223,29 +223,29 @@ EOM
 #!/bin/sh
 # $file Startup script for *nix systems
 
-if [ `expr "\$JAVA_OPTS" : ".*headless.*"` != 0 ]; then
-	# We look in \$JAVA_OPTS ... if we've done java.awt.headless mode 
-        # then we've likely got the default options for SSG and it would prevent a gui
-        # from coming up. So we over-write them with the following
-	JAVA_OPTS=" -Xms96M -Xmx96M -Xss256k -server -XX:NewSize=48M -XX:MaxNewSize=48M ";
-fi
 
-# if we don't have an L7_OPTS and we DO have a java opts, 
-# e.g. from above code or from the user's environment
-# we set the l7opts to be equal to java opts
-
-if [ "\$L7_OPTS" = "" -a "\$JAVA_OPTS" != "" ]; then
-	L7_OPTS=\$JAVA_OPTS
+if [ -z "\$JAVA_OPTS" ]; then
+	# we don't have java opts, so we set them ourselves
+	JAVA_OPTS=" -Xms128M -Xmx256M -Xss256k -server ";
+elif [ `expr "\$JAVA_OPTS" : ".*headless.*"` != 0 ]; then
+       # We look in \$JAVA_OPTS ... if java.awt.headless mode is there
+       # then we've likely got the default options for SSG and it would prevent a gui
+       # from coming up. So we over-write them with the following
+       JAVA_OPTS=" -Xms128M -Xmx256M -Xss256k -server ";
 fi
 
 # set current dir to where this script is
-
 cd `dirname \$0`
 
 # include startup options 
 $start_options
 
-\$JAVA_HOME/bin/java \$L7_OPTS \$extra \$run
+if [ -z "\$JAVA_HOME" ]; then
+        echo "No JAVA_HOME set, exiting"
+        exit;
+fi
+       
+\$JAVA_HOME/bin/java \$JAVA_OPTS \$extra \$run
 EOM
 
 	unlink("$dir.tar.gz");
